+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KK9GG464A5GRYXP9FR0W5H1K"
+        "01KKB0P285P7QMP52ZSKFHXY2V"
     )
     r"""The unique identifier of the data source"""