+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -64,7 +64,7 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-03-09T14:37:14.998Z")
+    created: Optional[datetime] = parse_datetime("2026-03-10T04:39:20.491Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")