+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KK9GG4615VZXFXN1DRZR7VF5"
+        "01KKB0P27AMVTRFP9HNZZVG6PT"
     )
     r"""The unique identifier of the data source"""