+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KK9GG463Y7MTVNR34T4SM44E"
+        "01KKB0P282CCM6VD05D20NFA7M"
     )
     r"""The unique identifier of the data source"""