+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -1794,7 +1794,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KK9GG3HQSHAFFZPWXTM6MKPS"
+    id: Optional[str] = "01KKB0P1E1YTNM09WWBR1NYS2N"
 
     description: Optional[str] = None